Replace Threads req/resp with info from trace
- make process and actor threads TracingActivityThreads, a new common super class
- make id part of Activity
- TraceData is supposed to encode/formalize basic properties of trace values
- include main actor into trace data
- added process creation/completion into trace data
- encoding of number values is designed so that they only use 53 significant bits, which fits into the JS integer values (i.e., the int part of a double)
- make sure symbol data is sent before trace data to resolve data dependencies (technically there is still a race between receiving those two…)
- handle race between activity and stack trace display by using a promise for each activity

+++ src/tools/debugger/frontend/ApplicationThreadStack.java
@@ -22,7 +22,7 @@ class ApplicationThreadStack {
    * direct access.
    */
   final ArrayList<Object> scopesAndObjects;
-  final HashMap<Object, Integer> scopesAndObjectsSet;
+  final HashMap<Object, Long> scopesAndObjectsSet;
 
   private final ArrayList<DebugStackFrame> stackFrames;
   private final SuspendedEvent event;
@@ -46,14 +46,14 @@ class ApplicationThreadStack {
     return stackFrames;
   }
 
-  int addScope(final MaterializedFrame frame,
+  long addScope(final MaterializedFrame frame,
       final MethodScope lexicalScope) {
     scopesAndObjects.add(new RuntimeScope(frame, lexicalScope));
     return getLastScopeOrVarId();
   }
 
-  int addObject(final Object obj) {
-    Integer idx = scopesAndObjectsSet.get(obj);
+  long addObject(final Object obj) {
+    Long idx = scopesAndObjectsSet.get(obj);
     if (idx == null) {
       scopesAndObjects.add(obj);
       idx = getLastScopeOrVarId();
@@ -68,7 +68,7 @@ class ApplicationThreadStack {
     return scopesAndObjects.get(localVarRef - 1);
   }
 
-  private int getLastScopeOrVarId() {
+  private long getLastScopeOrVarId() {
     // using size() means ids start with 1, which seems to be needed
     // otherwise, VS code ignores the top frame
     return suspension.getGlobalId(scopesAndObjects.size());
